feat: #220 리뷰 뷰 피쳐 연결
