+++ Spoony-iOS/Spoony-iOS/Source/Feature/MyPage/Review/ReviewsFeature.swift
@@ -1,0 +1,33 @@
+//
+//  ReviewsFeature.swift
+//  Spoony-iOS
+//
+//  Created by 이지훈 on 4/11/25.
+//
+
+import SwiftUI
+
+import ComposableArchitecture
+import TCACoordinators
+
+
+@Reducer
+struct ReviewsFeature {
+    @ObservableState
+    struct State: Equatable {
+        static let initialState = State()
+    }
+    
+    enum Action {
+        case routeToPreviousScreen
+    }
+    
+    var body: some ReducerOf<Self> {
+        Reduce { state, action in
+            switch action {
+            case .routeToPreviousScreen:
+                return .none
+            }
+        }
+    }
+}
